refactor(alerting): Use one file for provider implementation
